fix(dagml): feed models native float32 + storage-order refit rows — RF/GBR cross-engine parity (Codex order #2, Codex SHIP)
Codex order item #2. The conformance pack xfailed baseline_savgol_rf_kfold / baseline_detrend_firstderiv_gbr /
baseline_classification_rf_stratified as "fold-materialization divergence" — MISDIAGNOSED. Two host-side
(nirs4all, NOT dag-ml Rust) divergences at the model-fit data boundary, both real fidelity bugs:

* dtype: the dataset stores X as float32 (legacy contract) but the dag-ml host widened it to float64
  (resolver .tolist() + node_runner np.asarray(dtype=float)). Legacy feeds float32. SavGol(f32) vs (f64) differ
  ~8e-7 -> tips random_state=42 tree splits -> different trees. resolver.py preserves the x_rows NATIVE dtype;
  node_runner.py drops the dtype=float widening on fit/predict X (y stays float64). Now byte-faithful to legacy.
* refit row-order: envelope.py build_fold_set ordered the REFIT full-train sample_ids by fold-first-seen
  ([1,2,3,5,6,...]); legacy trains on the train partition in STORAGE order ([0,1,2,3,...]). Same set, different
  order -> fixed-seed RF/GBR bootstrap draws different rows. Now sorted(seen) = storage order. Per-fold
  train_sample_ids + OOF coverage unchanged (PLS/Ridge order-invariant -> no-op for them).

Validated Δ=0.0 on the 3 RF/GBR cases; flipped a 4th (generator_log_range_alpha 2.3e-3 -> 1.3e-7: the
ill-conditioned Ridge only amplified the float-widening noise). Full RF suite (conformance + export-roundtrip +
examples + cli_runner + node_runner + selector + native_results): 256 passed, 4 skipped, 11 xfailed, 0 failed.

The dtype pin made the engine correctly native-float32, so the test sklearn ORACLES (which computed expected in
float64) were aligned to native float32 (X only; y stays float64) — cli_runner (77 X-reads + 1 MB-PLS),
native_results (3), node_runner (2 SNV-chain), and the export-roundtrip test input (_test_x; _DagmlExportedModel
itself does NOT widen X — investigated, no export bug). No tolerance loosened; symmetric resolver-vs-resolver
checks untouched. ruff clean; mypy 0 new (8 pre-existing in cli_runner). Codex SHIP (3 rounds).

+++ nirs4all/pipeline/dagml/node_runner.py
@@ -384,13 +384,18 @@ def run_model_node(
         # wrapper applies the X-chain per block and fits ``model.fit([X1,X2,…], y)``. BY_SOURCE (S4):
         # materialize ONLY the bound source's block (one 2D matrix — late fusion by source). Otherwise the
         # single concat matrix (single-source OR early-fusion multi-source) is fit as before.
+        # X is materialized at the dataset's NATIVE storage dtype (no float64 widening): np.asarray on the
+        # resolver's ndarray is a dtype-preserving no-op (float32 for the legacy SpectroDataset contract).
+        # Legacy feeds the estimator float32; a float64 widening shifts inputs ~1e-7, which tips split
+        # thresholds on a fixed-seed tree ensemble (RF/GBR) and diverges the fitted trees. y stays float
+        # (legacy feeds y as float64).
         x_train: Any
         if multi_block:
-            x_train = [np.asarray(block, dtype=float) for block in resolver.resolve_feature_blocks(fit_ids, include_augmented=True)["blocks"]]
+            x_train = [np.asarray(block) for block in resolver.resolve_feature_blocks(fit_ids, include_augmented=True)["blocks"]]
         elif source_index is not None:
-            x_train = np.asarray(resolver.resolve_source_block(fit_ids, source_index, include_augmented=True)["values"], dtype=float)
+            x_train = np.asarray(resolver.resolve_source_block(fit_ids, source_index, include_augmented=True)["values"])
         else:
-            x_train = np.asarray(resolver.resolve_features(fit_ids, include_augmented=True)["values"], dtype=float)
+            x_train = np.asarray(resolver.resolve_features(fit_ids, include_augmented=True)["values"])
         y_train = np.asarray(resolver.resolve_targets(resolver.target_sample_ids(fit_ids))["values"], dtype=float)
         # MULTI-TARGET (S0): resolve_targets returns a 2D (n, n_targets) block, so y_train is already 2D
         # — pass it through unraveled (PLSRegression(n_targets>1)/MultiOutputRegressor consume 2D y) and
@@ -402,13 +407,15 @@ def run_model_node(
         estimator.fit(x_train, y_fit)
 
     def _predict(ids: list[str], include_augmented: bool) -> list[list[float]]:
+        # Predict X at the dataset's NATIVE storage dtype too (same parity reason as the fit X above):
+        # np.asarray on the resolver's ndarray preserves float32; legacy predicts on float32.
         x: Any
         if multi_block:
-            x = [np.asarray(block, dtype=float) for block in resolver.resolve_feature_blocks(ids, include_augmented=include_augmented)["blocks"]]
+            x = [np.asarray(block) for block in resolver.resolve_feature_blocks(ids, include_augmented=include_augmented)["blocks"]]
         elif source_index is not None:
-            x = np.asarray(resolver.resolve_source_block(ids, source_index, include_augmented=include_augmented)["values"], dtype=float)
+            x = np.asarray(resolver.resolve_source_block(ids, source_index, include_augmented=include_augmented)["values"])
         else:
-            x = np.asarray(resolver.resolve_features(ids, include_augmented=include_augmented)["values"], dtype=float)
+            x = np.asarray(resolver.resolve_features(ids, include_augmented=include_augmented)["values"])
         pred = np.asarray(estimator.predict(x), dtype=float).reshape(len(ids), -1)
         scaled = np.asarray(y_transform.inverse_transform(pred), dtype=float).reshape(len(ids), -1) if y_transform is not None else pred
         return [[float(value) for value in row] for row in scaled]
